+++ src/lib/barrels.ts
@@ -58,6 +58,7 @@ import { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module
 import { StWidgetModule } from './st-widget/st-widget.module';
 import { StFormFieldModule } from './st-form/st-form-field/st-form-field.module';
 import { StForegroundNotificationsModule } from './st-foreground-notifications/st-foreground-notifications.module';
+import { StZeroPageModule } from './st-zero-page/st-zero-page.module';
 
 export const DECLARATIONS: any[] = [
    PipesModule,
@@ -109,5 +110,6 @@ export const DECLARATIONS: any[] = [
    StTreeModule,
    StTwoListSelectionModule,
    StVerticalTabsModule,
-   StWidgetModule
+   StWidgetModule,
+   StZeroPageModule
 ];